+++ src/memoFix.mjs
@@ -1,15 +1,15 @@
 import { BoxedMultiKeyMap } from "./BoxedMultiKeyMap.mjs";
-import { Trait, apply, extend } from "./index.mjs";
+import { trait, apply, extend } from "./index.mjs";
 
 /**
  * Extends a trait to support the computation of a least fixed point with a bottom element and memoization
- * @param {Trait} trait The trait to extend
- * @param {any|() => any} bottom The bottom element of the least fixed point
- * @returns {Trait} The extended trait
+ * @param trait The trait to extend
+ * @param bottom The bottom element of the least fixed point
+ * @returns The extended trait
  */
 export const memoFix = (trait, bottom) => {
     const visited = new BoxedMultiKeyMap()
-    return Trait(undefined, {
+    return trait(undefined, {
         [extend]: trait,
         [apply](...args) {
             if (!visited.has(...args)) {