+++ src/tests/color.test.mjs
@@ -1,9 +1,9 @@
-import { Data, isData, data, variant, variantName, Trait } from "../index.mjs"
+import { data, isData, dataDecl, variant, variantName, trait } from "../index.mjs"
 
 describe('Color tests', () => {
-    const Color = Data({ Red: {}, Green: {}, Blue: {} });
+    const Color = data({ Red: {}, Green: {}, Blue: {} });
 
-    test('Color Data', () => {
+    test('Color data', () => {
         expect(Color[isData]).toBe(true);
         const red = Color.Red;
         expect(red).toBeDefined();
@@ -12,14 +12,14 @@ describe('Color tests', () => {
         expect(red.Red).toBeUndefined();
     })
 
-    const print = Trait(Color, {
+    const print = trait(Color, {
         Red() { return '#FF0000' },
         Green() { return '#00FF00' },
         Blue() { return '#0000FF' }
     })
 
     test('Color print', () => {
-        expect(print[data]).toBe(Color);
+        expect(print[dataDecl]).toBe(Color);
 
         expect(print(Color.Red)).toBe('#FF0000');
         expect(print(Color.Green)).toBe('#00FF00');