+++ src/tests/primitive-trait.test.mjs
@@ -1,8 +1,8 @@
-import { Trait, data } from "../index.mjs"
+import { trait, dataDecl } from "../index.mjs"
 
 describe('Primitive traits', () => {
     test('Number trait', () => {
-        const printNumber = Trait(Number, {
+        const printNumber = trait(Number, {
             1: (n) => 'one',
             15: (n) => 'fifteen',
             [Infinity]: (n) => 'infinity',
@@ -18,7 +18,7 @@ describe('Primitive traits', () => {
         })
 
         expect(printNumber).toBeDefined()
-        expect(printNumber[data]).toBe(Number)
+        expect(printNumber[dataDecl]).toBe(Number)
         expect(printNumber(1)).toBe('one')
         expect(printNumber(15)).toBe('fifteen')
         expect(printNumber(Infinity)).toBe('positive infinity')
@@ -37,14 +37,14 @@ describe('Primitive traits', () => {
     })
 
     test('String trait', () => {
-        const printString = Trait(String, {
+        const printString = trait(String, {
             '': (s) => 'empty string',
             'hello': (s) => s,
             _: (s) => s
         })
 
         expect(printString).toBeDefined()
-        expect(printString[data]).toBe(String)
+        expect(printString[dataDecl]).toBe(String)
         expect(printString('')).toBe('empty string')
         expect(printString('hello')).toBe('hello')
 
@@ -52,13 +52,13 @@ describe('Primitive traits', () => {
     })
 
     test('Boolean trait', () => {
-        const printBoolean = Trait(Boolean, {
+        const printBoolean = trait(Boolean, {
             true: () => 'true',
             false: () => 'false'
         })
 
         expect(printBoolean).toBeDefined()
-        expect(printBoolean[data]).toBe(Boolean)
+        expect(printBoolean[dataDecl]).toBe(Boolean)
         expect(printBoolean(true)).toBe('true')
         expect(printBoolean(false)).toBe('false')
 
@@ -66,7 +66,7 @@ describe('Primitive traits', () => {
     })
 
     test('BigInt trait', () => {
-        const printBigInt = Trait(BigInt, {
+        const printBigInt = trait(BigInt, {
             '0n': (n) => 'zero',
             '1n': (n) => 'one',
             '1234567890123456789012345678901234567890n': (n) => 'a big number',
@@ -74,7 +74,7 @@ describe('Primitive traits', () => {
         })
 
         expect(printBigInt).toBeDefined()
-        expect(printBigInt[data]).toBe(BigInt)
+        expect(printBigInt[dataDecl]).toBe(BigInt)
         expect(printBigInt(0n)).toBe('zero')
         expect(printBigInt(1n)).toBe('one')
         expect(printBigInt(1234567890123456789012345678901234567890n)).toBe('a big number')